pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "caretaker-qa"
version = "0.1.0"
description = "Security-relevance agent — pulls vuln feeds + security news, matches against a watchlist of public repos, surfaces actionable signals."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Ian Lintner" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.1,<9",
    "httpx>=0.27,<1",
    "pydantic>=2.8,<3",
    "pyyaml>=6,<7",
    "tenacity>=8.2,<10",
    "feedparser>=6.0,<7",
    "litellm>=1.50,<2",
    "langgraph>=0.2,<1",
    "rich>=13,<15",
    "packaging>=24,<25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-cov>=5,<7",
    "pytest-asyncio>=0.24,<1",
    "ruff>=0.7,<1",
    "mypy>=1.11,<2",
    "types-PyYAML",
    "respx>=0.21,<1",
]

[project.scripts]
qa-agent = "qa_agent.cli:main"

[project.urls]
Homepage = "https://github.com/ianlintner/caretaker-qa"
Issues = "https://github.com/ianlintner/caretaker-qa/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/qa_agent"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "B",    # bugbear
    "UP",   # pyupgrade
    "SIM",  # simplify
    "RUF",  # ruff-specific
]
ignore = ["E501"]  # line-length handled by formatter

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011", "SIM117"]

[tool.ruff.format]
line-ending = "lf"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["feedparser", "litellm", "langgraph.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
asyncio_mode = "auto"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:pydantic.*",
]

[tool.coverage.run]
source = ["src/qa_agent"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 80
